improve prediction code + layer updates

+++ code/apps/client/header/prediction.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "entity_view.h"
 
-float smooth_val(float cur, float tgt, float dt);
+float smooth_val(float cur, float tgt, uint64_t dt);
 void predict_receive_update(entity_view *d, entity_view *data);
